Updated pages and added sign in, sign up, and sign out
I changed the style up a bit and added a nav bar, I also added an auth_lib to handle the authentication for the sign in and sign up. I added the User class that helps with user sign in and sign out. I added files to update the nonprofits table and the users table.